refactor: move renderer-agnostic elements to core
