+++ packages/core/src/index.ts
@@ -3,3 +3,7 @@ export * from './layout/register-defaults.js';
 export * from './layout/rules/basic.js';
 export * from './layout/rules/extra.js';
 export * from './elements/UIElement.js';
+export * from './elements/BorderPanel.js';
+export * from './elements/Grid.js';
+export * from './elements/Image.js';
+export * from './elements/Text.js';
