+++ packages/parser/src/parsers/GridParser.ts
@@ -1,4 +1,5 @@
-import { Grid, Row, Col } from '../../../runtime/src/elements/Grid.js';
+import { Grid } from '../../../runtime/src/elements/Grid.js';
+import { Row, Col } from '../../../core/src/index.js';
 import { applyGridAttachedProps, parseLen, applyMargin } from '../../../runtime/src/helpers.js';
 import type { ElementParser } from './ElementParser.js';
 import type { Parser } from '../Parser.js';
