+++ packages/runtime/src/index.ts
@@ -1,5 +1,4 @@
 export * from './template.js';
-export { createDefaultRegistry, RuleRegistry } from '../../core/src/index.js';
 export * from './elements/BorderPanel.js';
 export * from './elements/Image.js';
 export * from './elements/Text.js';
